+++ src/utils/decorators/input-label/__spec__.js
@@ -84,8 +84,27 @@ class NamelessClass {
   };
 }
 
+
+class LabelHelpClass {
+  props = {
+    name: 'foo',
+    label: 'test label',
+    labelHelp: 'help label',
+    labelInline: true,
+    labelWidth: 20
+  };
+
+  context = {
+    form: {
+      model: 'model_1'
+    }
+  };
+}
+
 describe('InputLabel', () => {
-  let instanceBasic, instanceFalse, instanceUnNamed, instanceValidation, instanceAltValidation, instanceNameless;
+  let instanceBasic, instanceFalse, instanceUnNamed,
+      instanceValidation, instanceAltValidation, instanceNameless,
+      instanceLabelHelp;
 
   beforeEach(() => {
     let ExtendedClassOne = InputLabel(BasicClass);
@@ -105,6 +124,9 @@ describe('InputLabel', () => {
 
     let ExtendedClassSix = InputLabel(NamelessClass);
     instanceNameless = new ExtendedClassSix();
+
+    let ExtendedClassSeven = InputLabel(LabelHelpClass);
+    instanceLabelHelp = new ExtendedClassSeven();
   });
 
   describe('labelHTML', () => {
@@ -122,27 +144,27 @@ describe('InputLabel', () => {
 
     describe('when no label is provided', () => {
       it('titleizes the name to provide the label text', () => {
-        var label = instanceUnNamed.labelHTML;
+        let label = instanceUnNamed.labelHTML;
         expect(label.props.children).toEqual('Bar Qux');
       });
     });
 
     describe('when the input has a label', () => {
       it('sets the labelText to the passed in label', () => {
-        var label = instanceBasic.labelHTML;
+        let label = instanceBasic.labelHTML;
         expect(label.props.children).toEqual('test label');
       });
 
       describe('when the input has a validation with asterisk enabled', () => {
         it('adds additional symbols to the label', () => {
-          var label = instanceValidation.labelHTML;
+          let label = instanceValidation.labelHTML;
           expect(label.props.children).toEqual('Validate Label*');
         });
       });
 
       describe('when the input does not have a validation with asterisk enabled', () => {
         it('does not add additional symbols to the label', () => {
-          var label = instanceAltValidation.labelHTML;
+          let label = instanceAltValidation.labelHTML;
           expect(label.props.children).toEqual('Validate Label');
         });
       });
@@ -151,6 +173,22 @@ describe('InputLabel', () => {
     describe('when label width is passed', () => {
       it('sets a width for the label', () => {
         expect(instanceBasic.labelHTML.props.style.width).toEqual('20%');
+      });
+    });
+  });
+
+  describe('labelHelpHTML', () => {
+    describe('when label help is provided', () => {
+      it('renders the help within a span', () => {
+        let help = instanceLabelHelp.labelHelpHTML;
+        expect(help.type).toEqual('span');
+        expect(help.props.children).toEqual('help label');
+      });
+    });
+
+    describe('when label help is not provided', () => {
+      it('does not return a help span', () => {
+        expect(instanceBasic.labelHelpHTML).toBeUndefined();
       });
     });
   });