+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## Improvements
 
 * Improved store reset. `store.reset()` will now reset the store to its initial data, whether or not history is enabled.
+* Label can now have a subtitle. Pass `labelHelp='help message'` to any label enabled input
 * Add `tableHeader` prop to `Table` component that allows you to set a row wrapped in a `thead` tag.
+* Add `thead` prop to `Table` component that allows you to set a row wrapped in a `thead` tag.
 
 ## New Components
 
@@ -19,6 +21,7 @@
   * `border` - allows developers to disable border.
   * `padding` - allows developers to have control over padding size.
 
+* Message style has changed to follow toast style
 * Pill style has changed
 
 ## Improved Dialog
@@ -27,9 +30,11 @@
 
 ## Misc
 
+* Format i18n error number for numeric validation.
 * Allow Tables to shrink in size using the `shrink` prop.
 * Link component can now display with an icon.
 * Child components of Row can now use a `columnAlign` prop.
+* Toast onDismiss is now optional
 
 ## Bug Fixes